[modify] 회원가입 컴포넌트, reducer 함수, contextprovider 분리 #36
