+++ src/app/join/components/index.ts
@@ -1,0 +1,3 @@
+export { default as FirstStep } from "./FirstStep";
+export { default as SecondStep } from "./SecondStep";
+export { default as ThirdStep } from "./ThirdStep";
